Fixed multiple issues such as webmap not loading some territories data on startup, performances issues on player movements

+++ src/main/java/fr/rosstail/nodewar/storage/storagetype/SqlStorageRequest.java
@@ -46,11 +46,14 @@ public class SqlStorageRequest implements StorageRequest {
 
     @Override
     public void setupStorage(String host, short port, String database, String username, String password) {
-        // updaters
-        deleteTeamMemberDuplicate();
-        alterTeamMemberTable();
-        alterTerritoryTable();
-        //
+        if (doesTableExists(teamMemberTableName)) {
+            deleteTeamMemberDuplicate();
+            alterTeamMemberTable();
+        }
+
+        if (doesTableExists(territoryTableName)) {
+            alterTerritoryTable();
+        }
 
         createNodewarPlayerTable();
         createNodewarTeamTable();
@@ -58,6 +61,34 @@ public class SqlStorageRequest implements StorageRequest {
         createNodewarTeamRelationTable();
         createNodewarTerritoryTable();
         createNodewarBattlefieldTable();
+    }
+
+    public boolean doesTableExists(String tableName) {
+        boolean tableExists;
+        ResultSet rs = executeSQLQuery(openConnection(),
+                        "SELECT count(*)" +
+                                " FROM information_schema.tables" +
+                                " WHERE table_schema = DATABASE()" +
+                                " AND table_name = '" + tableName + "';");
+
+        try {
+            int count = rs.getInt(1);
+            tableExists = count > 0;
+            rs.close();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        } finally {
+            if (rs != null) {
+                try {
+                    if (!rs.isClosed()) {
+                        rs.close();
+                    }
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return tableExists;
     }
 
     public void createNodewarPlayerTable() {
@@ -90,16 +121,14 @@ public class SqlStorageRequest implements StorageRequest {
     public void deleteTeamMemberDuplicate() {
         String deleteDuplicatesRequest =
                 "DELETE FROM " + teamMemberTableName + " a"
-                + " WHERE EXISTS (SELECT " + teamMemberTableName + " b"
-                + " WHERE a.player_id = b.player_id"
-                + " AND a.id > b.id)";
-
+                        + " WHERE EXISTS (SELECT " + teamMemberTableName + " b"
+                        + " WHERE a.player_id = b.player_id"
+                        + " AND a.id > b.id)";
         executeSQLQuery(openConnection(), deleteDuplicatesRequest);
     }
 
     public void alterTeamMemberTable() {
         String setPlayerIdUnique = "ALTER TABLE " + teamMemberTableName + " ADD UNIQUE (player_id)";
-
         executeSQL(setPlayerIdUnique);
     }
 
@@ -191,6 +220,7 @@ public class SqlStorageRequest implements StorageRequest {
 
     @Override
     public boolean insertTeamModel(TeamModel model) {
+        System.out.println("SqlStorageRequest.insertTeamModel " + model.getName().toLowerCase() + " " + model.getShortName());
         String query = "INSERT INTO " + teamTableName + " (name, display, short, color, is_open, is_relation_open, is_permanent)"
                 + " VALUES (?, ?, ?, ?, ?, ?, ?);";
         String name = model.getName().toLowerCase();
@@ -619,10 +649,10 @@ public class SqlStorageRequest implements StorageRequest {
         String query = "UPDATE " + territoryTableName + " SET owner_team_id = ?, last_update = CURRENT_TIMESTAMP WHERE name = ?";
         try {
             String ownerName = model.getOwnerName();
-            NwITeam nwTeam = TeamManager.getManager().getStringTeamMap().get(ownerName);
-            if (nwTeam != null) {
+            NwITeam nwITeam = TeamManager.getManager().getStringTeamMap().get(ownerName);
+            if (nwITeam != null) {
                 executeSQLUpdate(query,
-                        nwTeam.getID(),
+                        nwITeam.getID(),
                         model.getName()
                 );
             } else {
